+++ actix/src/arbiter.rs
@@ -5,11 +5,12 @@ use std::thread;
 use futures::sync::oneshot::{channel, Sender};
 use futures::{future, Future, IntoFuture};
 use tokio::executor::current_thread::spawn;
-use tokio::runtime::current_thread::Runtime;
+use tokio::runtime::current_thread::Builder as RuntimeBuilder;
 use uuid::Uuid;
 
 use actor::Actor;
 use address::{channel, Addr, AddressReceiver};
+use clock::Clock;
 use context::Context;
 use handler::Handler;
 use mailbox::DEFAULT_CAPACITY;
@@ -27,29 +28,62 @@ thread_local!(
 
 /// Event loop controller
 ///
-/// Arbiter controls event loop in it's thread. Each arbiter runs in separate
+/// Arbiter controls event loop in its thread. Each arbiter runs in separate
 /// thread. Arbiter provides several api for event loop access. Each arbiter
 /// can belongs to specific `System` actor.
+///
+/// By default, a panic in an Arbiter does _not_ stop the rest of the System,
+/// unless the panic is in the System actor. Users of Arbiter can opt into
+/// shutting down the system on panic by using `Arbiter::builder()` and enabling
+/// `stop_system_on_panic`.
+#[derive(Debug)]
 pub struct Arbiter {
     stop: Option<Sender<i32>>,
+    stop_system_on_panic: bool,
 }
 
 impl Actor for Arbiter {
     type Context = Context<Self>;
 }
 
+impl Drop for Arbiter {
+    fn drop(&mut self) {
+        if self.stop_system_on_panic && thread::panicking() {
+            eprintln!("Panic in Arbiter thread, shutting down system.");
+            System::current().stop_with_code(1)
+        }
+    }
+}
+
 impl Arbiter {
     /// Spawn new thread and run event loop in spawned thread.
     /// Returns address of newly created arbiter.
-    pub fn new<T: Into<String>>(name: T) -> Addr<Arbiter> {
-        let (tx, rx) = std::sync::mpsc::channel();
+    /// Does not stop the system on panic.
+    pub fn builder() -> Builder {
+        Builder::new()
+    }
 
+    /// Spawn new thread and run event loop in spawned thread.
+    /// Returns address of newly created arbiter.
+    /// Does not stop the system on panic.
+    pub fn new<T: Into<String>>(name: T) -> Addr<Arbiter> {
+        Arbiter::new_with_builder(Arbiter::builder().name(name))
+    }
+
+    /// Spawn new thread and run event loop in spawned thread.
+    /// Returns address of newly created arbiter.
+    fn new_with_builder(mut builder: Builder) -> Addr<Arbiter> {
+        let (tx, rx) = std::sync::mpsc::channel();
         let id = Uuid::new_v4();
-        let name = format!("arbiter:{}:{}", id.hyphenated().to_string(), name.into());
+        let name = format!(
+            "arbiter:{}:{}",
+            id.to_hyphenated_ref().to_string(),
+            builder.name.take().unwrap_or_else(|| "actor".into())
+        );
         let sys = System::current();
 
         let _ = thread::Builder::new().name(name.clone()).spawn(move || {
-            let mut rt = Runtime::new().unwrap();
+            let mut rt = builder.runtime.build().unwrap();
 
             let (stop, stop_rx) = channel();
             NAME.with(|cell| *cell.borrow_mut() = Some(name));
@@ -59,17 +93,21 @@ impl Arbiter {
             System::set_current(sys);
 
             // start arbiter
-            let addr =
-                rt.block_on(future::lazy(move || {
-                    let addr = Actor::start(Arbiter { stop: Some(stop) });
+            let addr = rt
+                .block_on(future::lazy(move || {
+                    let addr = Actor::start(Arbiter {
+                        stop: Some(stop),
+                        stop_system_on_panic: builder.stop_system_on_panic,
+                    });
                     Ok::<_, ()>(addr)
                 })).unwrap();
             ADDR.with(|cell| *cell.borrow_mut() = Some(addr.clone()));
 
             // register arbiter
-            System::current()
-                .sys()
-                .do_send(RegisterArbiter(id.simple().to_string(), addr.clone()));
+            System::current().sys().do_send(RegisterArbiter(
+                id.to_simple_ref().to_string(),
+                addr.clone(),
+            ));
 
             if tx.send(addr).is_err() {
                 error!("Can not start Arbiter, remote side is dead");
@@ -84,7 +122,7 @@ impl Arbiter {
             // unregister arbiter
             System::current()
                 .sys()
-                .do_send(UnregisterArbiter(id.simple().to_string()));
+                .do_send(UnregisterArbiter(id.to_simple_ref().to_string()));
         });
 
         rx.recv().unwrap()
@@ -97,7 +135,10 @@ impl Arbiter {
 
         // start arbiter
         let ctx = Context::with_receiver(rx);
-        let fut = ctx.into_future(Arbiter { stop: None });
+        let fut = ctx.into_future(Arbiter {
+            stop: None,
+            stop_system_on_panic: true, // If the system Arbiter panics, stop the system.
+        });
         let addr = fut.address();
         Arbiter::spawn(fut);
         ADDR.with(|cell| *cell.borrow_mut() = Some(addr.clone()));
@@ -117,7 +158,7 @@ impl Arbiter {
         RUNNING.with(|cell| cell.set(false));
     }
 
-    /// Returns current arbiter's address
+    /// Returns current arbiter's name
     pub fn name() -> String {
         NAME.with(|cell| match *cell.borrow() {
             Some(ref name) => name.clone(),
@@ -171,10 +212,18 @@ impl Arbiter {
         A: Actor<Context = Context<A>>,
         F: FnOnce(&mut A::Context) -> A + Send + 'static,
     {
+        Arbiter::builder().start(f)
+    }
+
+    fn start_with_builder<A, F>(builder: Builder, f: F) -> Addr<A>
+    where
+        A: Actor<Context = Context<A>>,
+        F: FnOnce(&mut A::Context) -> A + Send + 'static,
+    {
         let (stx, srx) = channel::channel(DEFAULT_CAPACITY);
 
         // new arbiter
-        let addr = Arbiter::new("actor");
+        let addr = Arbiter::new_with_builder(builder);
 
         // create actor
         addr.do_send::<Execute>(Execute::new(move || {
@@ -215,5 +264,66 @@ impl<I: Send, E: Send> Handler<Execute<I, E>> for Arbiter {
 
     fn handle(&mut self, msg: Execute<I, E>, _: &mut Context<Self>) -> Result<I, E> {
         msg.exec()
+    }
+}
+
+/// Builder struct for an Arbiter.
+pub struct Builder {
+    /// Name of the Arbiter. Defaults to "actor" if unset.
+    name: Option<String>,
+
+    /// Whether the Arbiter will stop the whole System on uncaught panic. Defaults to false.
+    stop_system_on_panic: bool,
+
+    /// Tokio runtime builder.
+    runtime: RuntimeBuilder,
+}
+
+impl Builder {
+    fn new() -> Self {
+        Builder {
+            name: None,
+            stop_system_on_panic: false,
+            runtime: RuntimeBuilder::new(),
+        }
+    }
+
+    /// Sets the name of the Arbiter.
+    pub fn name<T: Into<String>>(mut self, name: T) -> Self {
+        self.name = Some(name.into());
+        self
+    }
+
+    /// Sets the option 'stop_system_on_panic' which controls whether the System is stopped when an
+    /// uncaught panic is thrown from an Actor in the Arbiter.
+    ///
+    /// Defaults to false.
+    pub fn stop_system_on_panic(mut self, stop_system_on_panic: bool) -> Self {
+        self.stop_system_on_panic = stop_system_on_panic;
+        self
+    }
+
+    /// Set the Clock instance that will be used by this Arbiter.
+    ///
+    /// Defaults to the clock used by the actix `System`, which defaults to the system clock.
+    pub fn clock(mut self, clock: Clock) -> Self {
+        self.runtime.clock(clock);
+        self
+    }
+
+    /// Spawn new thread and run event loop in spawned thread.
+    /// Returns address of newly created arbiter.
+    pub fn build(self) -> Addr<Arbiter> {
+        Arbiter::new_with_builder(self)
+    }
+
+    /// Start new arbiter and then start actor in created arbiter.
+    /// Returns `Addr<Syn, A>` of created actor.
+    pub fn start<A, F>(self, f: F) -> Addr<A>
+    where
+        A: Actor<Context = Context<A>>,
+        F: FnOnce(&mut A::Context) -> A + Send + 'static,
+    {
+        Arbiter::start_with_builder(self, f)
     }
 }